Remove last displayed card on panel restart
If last displayed card had no exit (or did not worked or did not
existed) and Callback was configured in a way that will show the last
card again and again, even a restart of the panel might not have solved
targeting this card. Now cache will be cleared on panel restart.